Move getV2ClientTSTypeRef to utils
