+++ src/transforms/v2-to-v3/ts-type/replaceTSTypeReference.ts
@@ -1,7 +1,6 @@
 import { Collection, Identifier, JSCodeshift, TSQualifiedName, TSTypeReference } from "jscodeshift";
 
-import { getV3ClientDefaultLocalName } from "../utils";
-import { getV2ClientTSTypeRef } from "./getV2ClientTSTypeRef";
+import { getV2ClientTSTypeRef, getV3ClientDefaultLocalName } from "../utils";
 import { getV2ClientTypeNames } from "./getV2ClientTypeNames";
 import { getV3ClientTypeName } from "./getV3ClientTypeName";
