+++ src/transforms/v2-to-v3/utils/index.ts
@@ -1,4 +1,5 @@
 export * from "./getV2ClientNewExpression";
+export * from "./getV2ClientTSTypeRef";
 export * from "./getV2ServiceModulePath";
 export * from "./getV3ClientDefaultLocalName";
 export * from "./isTypeScriptFile";
